requirements.txt
Flask
Flask-Session
flask-cors
openai
python-dotenv
redis
gunicorn

